+++ src/app/App.svelte
@@ -21,67 +21,6 @@
   import {isNil} from "ramda"
   import {onMount} from "svelte"
   import {memoize} from "src/util/misc"
-  import About from "src/app/views/About.svelte"
-  import Search from "src/app/views/Search.svelte"
-  import Bech32Entity from "src/app/views/Bech32Entity.svelte"
-  import Calendar from "src/app/views/Calendar.svelte"
-  import ChannelCreate from "src/app/views/ChannelCreate.svelte"
-  import ChannelsDetail from "src/app/views/ChannelsDetail.svelte"
-  import ChannelsList from "src/app/views/ChannelsList.svelte"
-  import DataExport from "src/app/views/DataExport.svelte"
-  import DataImport from "src/app/views/DataImport.svelte"
-  import EventDetail from "src/app/views/EventDetail.svelte"
-  import EventEdit from "src/app/views/EventEdit.svelte"
-  import EventDelete from "src/app/views/EventDelete.svelte"
-  import GroupList from "src/app/views/GroupList.svelte"
-  import GroupDetail from "src/app/views/GroupDetail.svelte"
-  import GroupCreate from "src/app/views/GroupCreate.svelte"
-  import GroupEdit from "src/app/views/GroupEdit.svelte"
-  import GroupInfo from "src/app/views/GroupInfo.svelte"
-  import GroupShare from "src/app/views/GroupShare.svelte"
-  import GroupRotate from "src/app/views/GroupRotate.svelte"
-  import Help from "src/app/views/Help.svelte"
-  import Home from "src/app/views/Home.svelte"
-  import InviteCreate from "src/app/views/InviteCreate.svelte"
-  import InviteAccept from "src/app/views/InviteAccept.svelte"
-  import LabelCreate from "src/app/views/LabelCreate.svelte"
-  import ListEdit from "src/app/views/ListEdit.svelte"
-  import ListList from "src/app/views/ListList.svelte"
-  import ListSelect from "src/app/views/ListSelect.svelte"
-  import ListingEdit from "src/app/views/ListingEdit.svelte"
-  import ListingDelete from "src/app/views/ListingDelete.svelte"
-  import Login from "src/app/views/Login.svelte"
-  import LoginBunker from "src/app/views/LoginBunker.svelte"
-  import LoginConnect from "src/app/views/LoginConnect.svelte"
-  import LoginPrivKey from "src/app/views/LoginPrivKey.svelte"
-  import LoginPubKey from "src/app/views/LoginPubKey.svelte"
-  import Logout from "src/app/views/Logout.svelte"
-  import Market from "src/app/views/Market.svelte"
-  import NoteCreate from "src/app/views/NoteCreate.svelte"
-  import NoteDetail from "src/app/views/NoteDetail.svelte"
-  import Notifications from "src/app/views/Notifications.svelte"
-  import Onboarding from "src/app/views/Onboarding.svelte"
-  import PersonDetail from "src/app/views/PersonDetail.svelte"
-  import PersonInfo from "src/app/views/PersonInfo.svelte"
-  import PersonFollows from "src/app/views/PersonFollows.svelte"
-  import PersonFollowers from "src/app/views/PersonFollowers.svelte"
-  import PersonList from "src/app/shared/PersonList.svelte"
-  import PublishInfo from "src/app/views/PublishInfo.svelte"
-  import QRCode from "src/app/views/QRCode.svelte"
-  import MediaDetail from "src/app/views/MediaDetail.svelte"
-  import RelayBrowse from "src/app/views/RelayBrowse.svelte"
-  import RelayDetail from "src/app/views/RelayDetail.svelte"
-  import RelayList from "src/app/views/RelayList.svelte"
-  import RelayReview from "src/app/views/RelayReview.svelte"
-  // import ReportCreate from "src/app/views/ReportCreate.svelte"
-  import ThreadDetail from "src/app/views/ThreadDetail.svelte"
-  import TopicFeed from "src/app/views/TopicFeed.svelte"
-  import UserContent from "src/app/views/UserContent.svelte"
-  import UserData from "src/app/views/UserData.svelte"
-  import UserKeys from "src/app/views/UserKeys.svelte"
-  import UserProfile from "src/app/views/UserProfile.svelte"
-  import UserSettings from "src/app/views/UserSettings.svelte"
-  import Zap from "src/app/views/Zap.svelte"
   import {logUsage} from "src/app/state"
   import {
     router,
@@ -100,63 +39,63 @@
 
   // Routes
 
-  router.register("/about", About)
-  router.register("/search", Search)
-  router.register("/events", Calendar)
+  router.register("/about", import("src/app/views/About.svelte"))
+  router.register("/search", import("src/app/views/Search.svelte"))
+  router.register("/events", import("src/app/views/Calendar.svelte"))
 
-  router.register("/channels", ChannelsList, {
+  router.register("/channels", import("src/app/views/ChannelsList.svelte"), {
     requireSigner: true,
   })
-  router.register("/channels/create", ChannelCreate, {
+  router.register("/channels/create", import("src/app/views/ChannelCreate.svelte"), {
     requireSigner: true,
   })
-  router.register("/channels/requests", ChannelsList, {
+  router.register("/channels/requests", import("src/app/views/ChannelsList.svelte"), {
     requireSigner: true,
   })
-  router.register("/channels/:channelId", ChannelsDetail, {
+  router.register("/channels/:channelId", import("src/app/views/ChannelsDetail.svelte"), {
     requireSigner: true,
     serializers: {
       channelId: asChannelId,
     },
   })
 
-  router.register("/events/:address", EventDetail, {
+  router.register("/events/:address", import("src/app/views/EventDetail.svelte"), {
     serializers: {
       address: asNaddr("address"),
     },
   })
-  router.register("/events/:address/edit", EventEdit, {
+  router.register("/events/:address/edit", import("src/app/views/EventEdit.svelte"), {
     serializers: {
       address: asNaddr("address"),
     },
   })
-  router.register("/events/:address/delete", EventDelete, {
+  router.register("/events/:address/delete", import("src/app/views/EventDelete.svelte"), {
     serializers: {
       address: asNaddr("address"),
     },
   })
 
-  router.register("/groups", GroupList)
-  router.register("/groups/new", GroupCreate, {
+  router.register("/groups", import("src/app/views/GroupList.svelte"))
+  router.register("/groups/new", import("src/app/views/GroupCreate.svelte"), {
     requireSigner: true,
   })
-  router.register("/groups/:address/edit", GroupEdit, {
+  router.register("/groups/:address/edit", import("src/app/views/GroupEdit.svelte"), {
     requireSigner: true,
     serializers: {
       address: asNaddr("address"),
     },
   })
-  router.register("/groups/:address/info", GroupInfo, {
+  router.register("/groups/:address/info", import("src/app/views/GroupInfo.svelte"), {
     serializers: {
       address: asNaddr("address"),
     },
   })
-  router.register("/groups/:address/share", GroupShare, {
+  router.register("/groups/:address/share", import("src/app/views/GroupShare.svelte"), {
     serializers: {
       address: asNaddr("address"),
     },
   })
-  router.register("/groups/:address/rotate", GroupRotate, {
+  router.register("/groups/:address/rotate", import("src/app/views/GroupRotate.svelte"), {
     requireSigner: true,
     serializers: {
       address: asNaddr("address"),
@@ -164,82 +103,82 @@
       removeMembers: asCsv("removeMembers"),
     },
   })
-  router.register("/groups/:address/:activeTab", GroupDetail, {
+  router.register("/groups/:address/:activeTab", import("src/app/views/GroupDetail.svelte"), {
     serializers: {
       address: asNaddr("address"),
     },
   })
 
-  router.register("/help/:topic", Help)
+  router.register("/help/:topic", import("src/app/views/Help.svelte"))
 
-  router.register("/invite", InviteAccept, {
+  router.register("/invite", import("src/app/views/InviteAccept.svelte"), {
     serializers: {
       people: asCsv("people"),
       relays: asCsv("relays"),
       groups: asCsv("groups"),
     },
   })
-  router.register("/invite/create", InviteCreate, {
+  router.register("/invite/create", import("src/app/views/InviteCreate.svelte"), {
     serializers: {
       initialPubkey: asUrlComponent("initialPubkey"),
       initialGroupAddress: asUrlComponent("initialGroupAddress"),
     },
   })
 
-  router.register("/lists", ListList)
-  router.register("/lists/create", ListEdit)
-  router.register("/lists/select", ListSelect, {
+  router.register("/lists", import("src/app/views/ListList.svelte"))
+  router.register("/lists/create", import("src/app/views/ListEdit.svelte"))
+  router.register("/lists/select", import("src/app/views/ListSelect.svelte"), {
     serializers: {
       type: asString("type"),
       value: asString("value"),
     },
   })
-  router.register("/lists/:address", ListEdit, {
+  router.register("/lists/:address", import("src/app/views/ListEdit.svelte"), {
     serializers: {
       address: asNaddr("address"),
     },
   })
 
-  router.register("/login", Login)
-  router.register("/login/bunker", LoginBunker)
-  router.register("/login/privkey", LoginPrivKey)
-  router.register("/login/pubkey", LoginPubKey)
-  router.register("/login/connect", LoginConnect, {
+  router.register("/login", import("src/app/views/Login.svelte"))
+  router.register("/login/bunker", import("src/app/views/LoginBunker.svelte"))
+  router.register("/login/privkey", import("src/app/views/LoginPrivKey.svelte"))
+  router.register("/login/pubkey", import("src/app/views/LoginPubKey.svelte"))
+  router.register("/login/connect", import("src/app/views/LoginConnect.svelte"), {
     requireUser: true,
   })
-  router.register("/logout", Logout)
+  router.register("/logout", import("src/app/views/Logout.svelte"))
 
-  router.register("/listings", Market)
-  router.register("/listings/:address/edit", ListingEdit, {
+  router.register("/listings", import("src/app/views/Market.svelte"))
+  router.register("/listings/:address/edit", import("src/app/views/ListingEdit.svelte"), {
     requireSigner: true,
     serializers: {
       address: asNaddr("address"),
     },
   })
-  router.register("/listings/:address/delete", ListingDelete, {
+  router.register("/listings/:address/delete", import("src/app/views/ListingDelete.svelte"), {
     requireSigner: true,
     serializers: {
       address: asNaddr("address"),
     },
   })
 
-  router.register("/media/:url", MediaDetail, {
+  router.register("/media/:url", import("src/app/views/MediaDetail.svelte"), {
     serializers: {
       url: asUrlComponent("url"),
     },
   })
 
-  router.register("/", Home, {
+  router.register("/", import("src/app/views/Home.svelte"), {
     serializers: {
       filter: asFilter,
     },
   })
-  router.register("/notes", Home, {
+  router.register("/notes", import("src/app/views/Home.svelte"), {
     serializers: {
       filter: asFilter,
     },
   })
-  router.register("/notes/create", NoteCreate, {
+  router.register("/notes/create", import("src/app/views/NoteCreate.svelte"), {
     requireSigner: true,
     serializers: {
       pubkey: asPerson,
@@ -247,112 +186,112 @@
       type: asString("type"),
     },
   })
-  router.register("/notes/:entity", NoteDetail, {
+  router.register("/notes/:entity", import("src/app/views/NoteDetail.svelte"), {
     serializers: {
       entity: asNote,
     },
   })
-  router.register("/notes/:entity/label", LabelCreate, {
+  router.register("/notes/:entity/label", import("src/app/views/LabelCreate.svelte"), {
     serializers: {
       entity: asNote,
     },
   })
-  router.register("/notes/:entity/status", PublishInfo, {
+  router.register("/notes/:entity/status", import("src/app/views/PublishInfo.svelte"), {
     serializers: {
       entity: asNote,
     },
   })
-  router.register("/notes/:entity/thread", ThreadDetail, {
+  router.register("/notes/:entity/thread", import("src/app/views/ThreadDetail.svelte"), {
     serializers: {
       entity: asNote,
     },
   })
 
-  router.register("/notifications", Notifications, {
+  router.register("/notifications", import("src/app/views/Notifications.svelte"), {
     requireUser: true,
   })
-  router.register("/notifications/:activeTab", Notifications, {
+  router.register("/notifications/:activeTab", import("src/app/views/Notifications.svelte"), {
     requireUser: true,
   })
 
-  router.register("/signup", Onboarding)
+  router.register("/signup", import("src/app/views/Onboarding.svelte"))
 
-  router.register("/people/list", PersonList, {
+  router.register("/people/list", import("src/app/shared/PersonList.svelte"), {
     serializers: {
       pubkeys: asCsv("pubkeys"),
     },
   })
-  router.register("/people/:entity", PersonDetail, {
+  router.register("/people/:entity", import("src/app/views/PersonDetail.svelte"), {
     required: ["pubkey"],
     serializers: {
       entity: asPerson,
       filter: asFilter,
     },
   })
-  router.register("/people/:entity/followers", PersonFollowers, {
+  router.register("/people/:entity/followers", import("src/app/views/PersonFollowers.svelte"), {
     required: ["pubkey"],
     serializers: {
       entity: asPerson,
     },
   })
-  router.register("/people/:entity/follows", PersonFollows, {
+  router.register("/people/:entity/follows", import("src/app/views/PersonFollows.svelte"), {
     required: ["pubkey"],
     serializers: {
       entity: asPerson,
     },
   })
-  router.register("/people/:entity/info", PersonInfo, {
+  router.register("/people/:entity/info", import("src/app/views/PersonInfo.svelte"), {
     required: ["pubkey"],
     serializers: {
       entity: asPerson,
     },
   })
 
-  router.register("/qrcode/:code", QRCode, {
+  router.register("/qrcode/:code", import("src/app/views/QRCode.svelte"), {
     serializers: {
       code: asUrlComponent("code"),
     },
   })
 
-  router.register("/relays/browse", RelayBrowse)
-  router.register("/relays/:entity", RelayDetail, {
+  router.register("/relays/browse", import("src/app/views/RelayBrowse.svelte"))
+  router.register("/relays/:entity", import("src/app/views/RelayDetail.svelte"), {
     serializers: {
       entity: asRelay,
       filter: asFilter,
     },
   })
-  router.register("/relays/:entity/review", RelayReview, {
+  router.register("/relays/:entity/review", import("src/app/views/RelayReview.svelte"), {
     serializers: {
       entity: asRelay,
     },
   })
 
-  router.register("/settings", UserSettings, {
+  router.register("/settings", import("src/app/views/UserSettings.svelte"), {
     requireUser: true,
   })
-  router.register("/settings/content", UserContent, {
+  router.register("/settings/content", import("src/app/views/UserContent.svelte"), {
     requireUser: true,
   })
-  router.register("/settings/data", UserData, {
+  router.register("/settings/data", import("src/app/views/UserData.svelte"), {
     requireUser: true,
   })
-  router.register("/settings/data/export", DataExport, {
+  router.register("/settings/data/export", import("src/app/views/DataExport.svelte"), {
     requireUser: true,
   })
-  router.register("/settings/data/import", DataImport, {
+  router.register("/settings/data/import", import("src/app/views/DataImport.svelte"), {
     requireUser: true,
   })
-  router.register("/settings/keys", UserKeys, {
+  router.register("/settings/keys", import("src/app/views/UserKeys.svelte"), {
     requireUser: true,
   })
-  router.register("/settings/profile", UserProfile, {
+  router.register("/settings/profile", import("src/app/views/UserProfile.svelte"), {
     requireUser: true,
   })
-  router.register("/settings/relays", RelayList)
+  router.register("/settings/relays", import("src/app/views/RelayList.svelte"))
 
-  router.register("/topics/:topic", TopicFeed)
+  router.register("/topics/:topic", import("src/app/views/TopicFeed.svelte"))
 
-  router.register("/zap", Zap, {
+  router.register("/zap", import("src/app/views/Zap.svelte"), {
     required: ["splits"],
     serializers: {
       eid: asNote,
@@ -361,13 +300,13 @@
     },
   })
 
-  router.register("/:entity", Bech32Entity, {
+  router.register("/:entity", import("src/app/views/Bech32Entity.svelte"), {
     serializers: {
       entity: asEntity,
       filter: asFilter,
     },
   })
-  router.register("/:entity/*", Bech32Entity, {
+  router.register("/:entity/*", import("src/app/views/Bech32Entity.svelte"), {
     serializers: {
       entity: asEntity,
       filter: asFilter,