+++ src/app/MenuDesktop.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import {toggleTheme, theme, appName} from "src/partials/state"
+  import {toggleTheme, theme} from "src/partials/state"
   import MenuItem from "src/partials/MenuItem.svelte"
   import FlexColumn from "src/partials/FlexColumn.svelte"
   import Anchor from "src/partials/Anchor.svelte"
@@ -56,11 +56,9 @@
     href="https://info.coracle.social">
     <img
       alt="App Logo"
-      src={
-        $theme === "dark"
-          ? import.meta.env.VITE_APP_WORDMARK_DARK
-          : import.meta.env.VITE_APP_WORDMARK_LIGHT
-      } />
+      src={$theme === "dark"
+        ? import.meta.env.VITE_APP_WORDMARK_DARK
+        : import.meta.env.VITE_APP_WORDMARK_LIGHT} />
   </Anchor>
   <MenuDesktopItem path="/notes">Feed</MenuDesktopItem>
   {#if !$env.FORCE_GROUP && $env.PLATFORM_RELAYS.length === 0}